new stuff
Former-commit-id: 1266ae5e1705f5d753ef57573cdc93ca2728cb04

+++ app/Http/Controllers/Advertiser/CampaignsController.php
@@ -139,7 +139,17 @@ class CampaignsController extends Controller
         ]);
 
         if($request->min_age < 0 || $request->max_age < 0){
-            return back()->withErrors('The minimun or maximum age cannot assume a negetive value');
+            Session::flash('error', 'The minimun or maximum age cannot assume a negetive value');
+            return back();
+        }
+
+        if($request->min_age > $request->max_age){
+            Session::flash('error', 'The minimum age cannot be greater than the maximum age');
+            return back();
+        }
+
+        if(strtotime($request->end_date) < strtotime($request->start_date)){
+            return redirect()->back()->with('error', 'Start Date cannot be greater than End Date');
         }
 
         $del_cart = \DB::delete("DELETE FROM carts WHERE user_id = '$id'");
